pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "consumed-ai"
version = "0.1.0"
description = "Universal software execution engine — execute any API via natural language"
readme = "README.md"
license = {text = "Apache-2.0"}
requires-python = ">=3.10"
authors = [{name = "Automate My", email = "hello@consumed.ai"}]
keywords = ["ai", "automation", "api", "agent", "llm", "chat"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: Apache Software License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Software Development :: Libraries :: Application Frameworks",
]

dependencies = [
    "fastapi>=0.110.0",
    "uvicorn[standard]>=0.27.0",
    "httpx>=0.27.0",
    "click>=8.1.0",
    "pydantic>=2.5.0",
    "pydantic-settings>=2.1.0",
    "cryptography>=42.0.0",
    "redis>=5.0.0",
    "pyyaml>=6.0",
    "rich>=13.0.0",
]

[project.optional-dependencies]
telegram = ["httpx>=0.27.0"]
discord = ["discord.py>=2.3.0"]
slack = ["slack-sdk>=3.27.0"]
all-channels = ["discord.py>=2.3.0", "slack-sdk>=3.27.0"]

[project.scripts]
consumed-ai = "consumed_ai.cli:main"

[project.urls]
Homepage = "https://consumed.ai"
Repository = "https://github.com/automatemy/consumed-ai"
Documentation = "https://docs.consumed.ai"
